backend/foodgram/requirements.txt
#autopep8==2.0.2
django-extra-fields==3.0.2
django-filter==23.1
djoser==2.1.0
flake8==5.0.4
#isort==5.11.5
#pip-chill==1.0.3
psycopg2-binary==2.9.6
python-dotenv==0.21.1
reportlab==3.6.12
gunicorn==20.0.4
django-extensions